refactor: move LdtkLevel logic to its own module
