+++ src/assets/mod.rs
@@ -1,72 +1,14 @@
 //! Assets and AssetLoaders for loading ldtk files.
 
-use crate::ldtk;
-use bevy::{
-    asset::{AssetLoader, AssetPath, LoadContext, LoadedAsset},
-    prelude::*,
-    reflect::TypeUuid,
-    utils::BoxedFuture,
-};
+use bevy::asset::AssetPath;
 use std::path::Path;
+
+mod ldtk_level;
+pub use ldtk_level::{LdtkLevel, LdtkLevelLoader};
 
 mod ldtk_project;
 pub use ldtk_project::{LdtkProject, LdtkProjectLoader};
 
 fn ldtk_path_to_asset_path<'b>(ldtk_path: &Path, rel_path: &str) -> AssetPath<'b> {
     ldtk_path.parent().unwrap().join(Path::new(rel_path)).into()
-}
-
-/// Secondary asset for loading ldtk files, specific to level data.
-///
-/// Loaded as a labeled asset when loading a standalone ldtk file with [LdtkProject].
-/// The label is just the level's identifier.
-///
-/// Loaded as a dependency to the [LdtkProject] when loading an ldtk file with external levels.
-#[derive(TypeUuid, Reflect, FromReflect)]
-#[uuid = "5448469b-2134-44f5-a86c-a7b829f70a0c"]
-pub struct LdtkLevel {
-    pub level: ldtk::Level,
-    pub background_image: Option<Handle<Image>>,
-}
-
-#[derive(Copy, Clone, Debug, Default)]
-pub struct LdtkLevelLoader;
-
-impl AssetLoader for LdtkLevelLoader {
-    fn load<'a>(
-        &'a self,
-        bytes: &'a [u8],
-        load_context: &'a mut LoadContext,
-    ) -> BoxedFuture<'a, anyhow::Result<()>> {
-        Box::pin(async move {
-            let level: ldtk::Level = serde_json::from_slice(bytes)?;
-
-            let mut background_asset_path = None;
-            let mut background_image = None;
-            if let Some(rel_path) = &level.bg_rel_path {
-                let asset_path =
-                    ldtk_path_to_asset_path(load_context.path().parent().unwrap(), rel_path);
-                background_asset_path = Some(asset_path.clone());
-                background_image = Some(load_context.get_handle(asset_path));
-            }
-
-            let ldtk_level = LdtkLevel {
-                level,
-                background_image,
-            };
-
-            let mut loaded_asset = LoadedAsset::new(ldtk_level);
-
-            if let Some(asset_path) = background_asset_path {
-                loaded_asset = loaded_asset.with_dependency(asset_path);
-            }
-
-            load_context.set_default_asset(loaded_asset);
-            Ok(())
-        })
-    }
-
-    fn extensions(&self) -> &[&str] {
-        &["ldtkl"]
-    }
 }
